refactor: move assessmentplan to oscal package and remove unused test helpers
