+++ internal/api/handler/oscal/api.go
@@ -2,7 +2,6 @@ package oscal
 
 import (
 	"github.com/compliance-framework/configuration-service/internal/api"
-	assessmentplan "github.com/compliance-framework/configuration-service/internal/api/handler/oscal/assessmentplan"
 	"github.com/compliance-framework/configuration-service/internal/api/middleware"
 	"github.com/compliance-framework/configuration-service/internal/config"
 	"go.uber.org/zap"
@@ -34,6 +33,6 @@ func RegisterHandlers(server *api.Server, logger *zap.SugaredLogger, db *gorm.DB
 	poamHandler := NewPlanOfActionAndMilestonesHandler(logger, db)
 	poamHandler.Register(oscalGroup.Group("/plan-of-action-and-milestones"))
 
-	assessmentPlanHandler := assessmentplan.NewAssessmentPlanHandler(logger, db)
+	assessmentPlanHandler := NewAssessmentPlanHandler(logger, db)
 	assessmentPlanHandler.Register(oscalGroup.Group("/assessment-plans"))
 }
