+++ internal/api/handler/oscal/asset.go
@@ -1,4 +1,4 @@
-package assessmentplan
+package oscal
 
 import (
 	"fmt"
