+++ internal/api/handler/oscal/subject.go
@@ -1,4 +1,4 @@
-package assessmentplan
+package oscal
 
 import (
 	"fmt"
