+++ internal/api/handler/oscal/subject_integration_test.go
@@ -1,6 +1,6 @@
 //go:build integration
 
-package assessmentplan
+package oscal
 
 import (
 	"bytes"
